[FIX] product_variant_configurator: Fix purchase_variant_configurator_on_confirm tests
We need to add this for avoiding an odoo.exceptions.AccessError due
to some refactoring done upstream on read method + variant name_get
in Odoo. With this, we avoid to call super on the specific case of
virtual records, providing simply the name, which is acceptable.

This was used in `purchase_variant_configurator_on_confirm` module for increasing
inheritability playing the purchase.order.line product onchange on a virtual
record.